refactor(machine): move reboot related function in its own file
Before, all reboot related function where in state.go. This is not consistent with other
functionalities.
After, reboot related function and tests are in their own file, to keep file short in length.

I have also fixed few receiver naming which wasn't consistent over the package.
